Fix CheckStyle
