+++ src/main/java/seedu/address/ui/ModuleCard.java
@@ -4,7 +4,6 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.module.Grade;
 import seedu.address.model.module.Module;
 
 /**
